+++ redwood-composeui/src/commonMain/kotlin/app/cash/redwood/composeui/RedwoodContent.kt
@@ -29,8 +29,11 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.platform.LocalLayoutDirection
+import androidx.compose.ui.unit.LayoutDirection
 import app.cash.redwood.compose.RedwoodComposition
 import app.cash.redwood.ui.Density
+import app.cash.redwood.ui.LayoutDirection as RedwoodLayoutDirection
 import app.cash.redwood.ui.OnBackPressedDispatcher
 import app.cash.redwood.ui.Size
 import app.cash.redwood.ui.UiConfiguration
@@ -55,6 +58,10 @@ public fun RedwoodContent(
     safeAreaInsets = safeAreaInsets(),
     viewportSize = viewportSize,
     density = density.density.toDouble(),
+    layoutDirection = when (LocalLayoutDirection.current) {
+      LayoutDirection.Ltr -> RedwoodLayoutDirection.Ltr
+      LayoutDirection.Rtl -> RedwoodLayoutDirection.Rtl
+    },
   )
   val uiConfigurations = remember { MutableStateFlow(uiConfiguration) }
   LaunchedEffect(uiConfiguration) {